# api-spec.yaml
openapi: 3.0.3
info:
  title: Grants Query & Referral – Frontend API
  version: 0.4.1
  description: |
    Frontend-facing API for templates, chat sessions, uploads, analytics, and admin config.
    Flow: create chat → upload files → upsert answers (common→simple/complex) → finalize (AI or human).

    NOTE: ID fields are unified — the chat/session identifier is consistently named **chat_id**
    across all endpoints (including escalation and legacy submission).
servers:
  - url: https://api.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KPIResponse:
      type: object
      properties:
        total_queries: { type: integer, example: 100 }
        simple_queries: { type: integer, example: 70 }
        ai_resolved_queries: { type: integer, example: 50 }
    SingleChoiceAnalyticsRequest:
      type: object
      required: [start_time, end_time]
      properties:
        start_time: { type: string, example: "01/09/2025", description: dd/MM/yyyy }
        end_time:   { type: string, example: "01/10/2025", description: dd/MM/yyyy }
    SingleChoiceAnalyticsItem:
      type: object
      properties:
        question: { type: string, example: What is your role? }
        options:
          type: array
          items:
            type: object
            properties:
              label: { type: string, example: Researcher }
              count: { type: integer, example: 10 }
...
    ChoiceAnalyticsRequest:
      type: object
      required: [start_time, end_time]
      properties:
        start_time: { type: string, example: "01/09/2025", description: dd/MM/yyyy }
        end_time:   { type: string, example: "01/10/2025", description: dd/MM/yyyy }
        template:
          type: string
          nullable: true
          description: Optional filter for template type
          enum: [common, simple, complex]
    ChoiceAnalyticsItem:
      type: object
      properties:
        question: { type: string, example: "What is your role?" }
        type:
          type: string
          enum: [single, multi]
          example: single
        options:
          type: array
          items:
            type: object
            properties:
              label: { type: string, example: "Researcher" }
              count: { type: integer, example: 42 }
paths:
...
  /kpis:
    get:
      tags: [Analytics]
      summary: Get KPI information
      description: Returns aggregate KPIs. (Auth required)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: start_time
          schema: { type: string, example: "01/09/2025" }
        - in: query
          name: end_time
          schema: { type: string, example: "01/10/2025" }
      responses:
        '200':
          description: KPI summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KPIResponse' }
              examples:
                sample:
                  value: { total_queries: 120, simple_queries: 80, ai_resolved_queries: 60 }
        '401': { description: Unauthorized }

  # ---------- NEW: Unified choice analytics (single + multi) ----------
  /analytics/choice:
    post:
      tags: [Analytics]
      summary: Choice question distribution (single + multi)
      description: |
        Returns option selection counts for both single- and multi-choice questions
        between dates. (Auth required)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChoiceAnalyticsRequest' }
            examples:
              range:
                value: { start_time: "01/09/2025", end_time: "01/10/2025", template: "common" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChoiceAnalyticsItem' }
              examples:
                sample:
                  value:
                    - question: "What is your role?"
                      type: "single"
                      options:
                        - { label: "Researcher", count: 42 }
                        - { label: "Sponsor",   count: 18 }
                    - question: "Which topics apply?"
                      type: "multi"
                      options:
                        - { label: "Indemnity", count: 27 }
                        - { label: "IP",        count: 31 }
        '401': { description: Unauthorized }

  # ---------- NEW: Excel export for choice analytics ----------
  /analytics/choice-export:
    get:
      tags: [Analytics]
      summary: Export choice analytics as Excel (.xlsx)
      description: |
        Returns an Excel file containing choice analytics (single + multi) for the given range.
        (Auth required)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: start_time
          required: true
          schema: { type: string, example: "01/09/2025" }
          description: dd/MM/yyyy
        - in: query
          name: end_time
          required: true
          schema: { type: string, example: "01/10/2025" }
          description: dd/MM/yyyy
        - in: query
          name: template
          required: false
          schema: { type: string, enum: [common, simple, complex] }
          description: Optional template filter
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401': { description: Unauthorized }

  # ---------- (Deprecated) Single-choice analytics (kept for BC) ----------
  /analytics/single-choice:
    post:
      deprecated: true
      tags: [Analytics]
      summary: (Deprecated) Single-choice distribution across templates
      description: Replaced by POST /analytics/choice (covers single + multi). (Auth required)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SingleChoiceAnalyticsRequest' }
            examples:
              range:
                value: { start_time: "01/09/2025", end_time: "01/10/2025" }
      responses:
        '200':
          description: OK (use /analytics/choice going forward)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SingleChoiceAnalyticsItem' }
              examples:
                sample:
                  value:
                    - question: "What is your role?"
                      options:
                        - { label: "Researcher", count: 42 }
                        - { label: "Sponsor",   count: 18 }
        '401': { description: Unauthorized }

  # ---------- Admin / Config ----------
...